linux: dynamically load libX11
Removing linking of libX11 and replacing it with dlopen to call the
functions dynamically.
For the future wayland support, the library will not require the libX11
library to be installed on the running system.
For compiling it we need it, because of the huge and complicated type
definitions, but the compiled binary will run also on a system without
libX11.

References: https://github.com/golang-design/clipboard/issues/6
Signed-off-by: Vladimír Magyar <vladimir@mgyar.me>